requirements.txt
grpcio==1.41.0
grpcio-tools==1.41.0
numpy==1.21.4
